# WIP feat: move mini to standalone berty-mini binary