--- a/go/gen.sum
+++ b/go/gen.sum
@@ -4,4 +4,4 @@ cdcfd4235af64debca29d0553209a34df7899361  ../api/bertytypes.proto
 f8f05c10a1cf5c8b42f9bde6bb572d0bc687979e  ../api/errcode.proto
 5589d560e33f2da4a466ad965eb9c8bd3d7612cd  ../api/go-internal/handshake.proto
 6708726752b27f538549fe0c30b8f73f7e3574a5  ../api/go-internal/records.proto
-0b070bad0039128c90bc5f988f1eb0154b70899a  Makefile
+625f17f6175ca66a2650347582a194531493b99e  Makefile