--- a/tool/berty-mini-cluster/commander
+++ b/tool/berty-mini-cluster/commander
@@ -32,8 +32,8 @@ trap cleanup EXIT
 # the tmux
 TMUX=                                                                                                                      `# force TMUX to run, even if nested` \
     tmux \
-    new-session "source .env && berty mini -g \$GRP_KEY; read" \; select-pane -T no-docker \;                              `# run a simple instance` \
-    split-window "source .env && docker run -it --rm bertytech/berty mini -g \$GRP_KEY; read" \; select-pane -T docker \;  `# run an instance within docker` \
+    new-session "source .env && berty-mini -g \$GRP_KEY; read" \; select-pane -T no-docker \;                              `# run a simple instance` \
+    split-window "source .env && docker run -it --rm bertytech/berty-mini -g \$GRP_KEY; read" \; select-pane -T docker \;  `# run an instance within docker` \
     split-window "docker-compose run aaron; read" \; select-pane -T compose-aaron \;                                       `# run instances within docker-compose` \
     split-window "docker-compose run betty; read" \; select-pane -T compose-betty \; \
     split-window "docker-compose run chris; read" \; select-pane -T compose-chris \; \