# experiment:

Expected:
* entry points easier to maintain
* speed-up `go install ./cmd/only-one-command` for people making a lot of changes on just a specific aspect of the project
* smaller individual binaries (sum will be higher for sure) 
* having a binary that is more "external" -> example for people implementing small bots/clients

## before

```console
$ time make install
GO111MODULE=on go install -ldflags="-X berty.tech/berty/v2/go/pkg/bertymessenger.VcsRef=`git rev-parse --short HEAD` -X berty.tech/berty/v2/go/pkg/bertymessenger.Version=`git describe --tags --always` -X berty.tech/berty/v2/go/pkg/bertymessenger.BuildTime=`date +%s`" -v ./cmd/...

real    0m8,938s
user    0m16,184s
sys     0m2,051s
```

```console
$ ls -lah ~/go/bin/ | grep -E '(rdvp|berty$|berty-mini$)'
-rwxrwxr-x 1 moul moul  72M juil. 18 12:41 berty
-rwxrwxr-x 1 moul moul  27M juil. 18 12:41 rdvp
```

## After

```console
$ clear; time make go.install.berty-mini
GO111MODULE=on go install -ldflags="-X berty.tech/berty/v2/go/pkg/bertymessenger.VcsRef=`git rev-parse --short HEAD` -X berty.tech/berty/v2/go/pkg/bertymessenger.Version=`git describe --tags --always` -X berty.tech/berty/v2/go/pkg/bertymessenger.BuildTime=`date +%s`" -v ./cmd/berty-mini

real    0m8,995s
user    0m11,810s
sys     0m1,939s
$ time make install
GO111MODULE=on go install -ldflags="-X berty.tech/berty/v2/go/pkg/bertymessenger.VcsRef=`git rev-parse --short HEAD` -X berty.tech/berty/v2/go/pkg/bertymessenger.Version=`git describe --tags --always` -X berty.tech/berty/v2/go/pkg/bertymessenger.BuildTime=`date +%s`" -v ./cmd/...

real    0m9,340s
user    0m24,137s
sys     0m2,869s

$ clear; time make go.install.berty
GO111MODULE=on go install -ldflags="-X berty.tech/berty/v2/go/pkg/bertymessenger.VcsRef=`git rev-parse --short HEAD` -X berty.tech/berty/v2/go/pkg/bertymessenger.Version=`git describe --tags --always` -X berty.tech/berty/v2/go/pkg/bertymessenger.BuildTime=`date +%s`" -v ./cmd/berty

real    0m8,461s
user    0m11,883s
sys     0m1,567s
```

```console
$ ls -lah ~/go/bin/ | grep -E '(rdvp|berty$|berty-mini$)'
-rwxrwxr-x 1 moul moul  71M juil. 20 17:03 berty
-rwxrwxr-x 1 moul moul  70M juil. 20 17:02 berty-mini
-rwxrwxr-x 1 moul moul  27M juil. 20 17:02 rdvp
```